Save renders on its own module
